+++ fsspec/implementations/reference.py
@@ -276,8 +276,6 @@ class ReferenceFileSystem(AsyncFileSystem):
 
         @lru_cache(1000)
         def _render_jinja(u):
-            import jinja2
-
             return jinja2.Template(u).render(**self.templates)
 
         for k, v in references.get("refs", {}).items():
@@ -288,7 +286,6 @@ class ReferenceFileSystem(AsyncFileSystem):
             else:
                 u = v[0]
                 if "{{" in u:
-                    u = jinja2.Template(u).render(**self.templates)
                     if self.simple_templates:
                         u = (
                             u.replace("{{", "{")