+++ docs/source/changelog.rst
@@ -6,7 +6,7 @@ Dev
 
 Enhancements
 
-- Introduce ``fsspec.asyn.fsspec_loop`` to temporarily switch to the fsspec loop.
+- Introduce ``fsspec.asyn.fsspec_loop`` to temporarily switch to the fsspec loop. (#671)
 
 
 2021.06.0